# src/api/v1/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nubabel Public API
  description: |
    Public API for Nubabel - Multi-Agent Orchestration Platform.

    ## Authentication
    All endpoints require Bearer token authentication using an API key.
    Include the header: `Authorization: Bearer <your_api_key>`

    ## Rate Limiting
    Rate limits depend on your API key tier:
    - **Free**: 60 requests/minute, 1,000 requests/day
    - **Pro**: 300 requests/minute, 10,000 requests/day
    - **Enterprise**: 1,000 requests/minute, 100,000 requests/day

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit-Minute`: Your per-minute limit
    - `X-RateLimit-Remaining-Minute`: Remaining requests this minute
    - `X-RateLimit-Limit-Day`: Your per-day limit
    - `X-RateLimit-Remaining-Day`: Remaining requests today

    ## Errors
    All errors follow a consistent format:
    ```json
    {
      "error": "error_code",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Nubabel Support
    email: support@nubabel.io
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Production API
...
security:
  - BearerAuth: []
paths:
...
  /webhooks:
    get:
      summary: List webhooks
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      availableEvents:
                        type: array
                        items:
                          type: string

    post:
      summary: Create webhook
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/WebhookEvent'
      responses:
        '201':
          description: Webhook created (secret only shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Webhook'
                      - type: object
                        properties:
                          secret:
                            type: string
                            description: Webhook signing secret (only shown on creation)
                  message:
                    type: string
  /webhooks/{webhookId}:
    get:
      summary: Get webhook details
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Webhook'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update webhook
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/WebhookEvent'
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Webhook'

    delete:
      summary: Delete webhook
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Webhook deleted
  /webhooks/{webhookId}/test:
    post:
      summary: Test webhook
      description: Send a test payload to verify webhook configuration
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                      status:
                        type: integer
                      error:
                        type: string
                      latencyMs:
                        type: number
  /webhooks/{webhookId}/rotate-secret:
    post:
      summary: Rotate webhook secret
      description: Generate a new signing secret (old secret immediately invalidated)
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New secret generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      secret:
                        type: string
                  message:
                    type: string
...
  /webhooks/events/list:
    get:
      summary: List available webhook events
      tags: [Webhooks]
      security:
        - BearerAuth: [webhooks:manage]
      responses:
        '200':
          description: Available events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        description:
                          type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the Developer Portal
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        status:
          type: string
          enum: [active, disabled]
        failureCount:
          type: integer
        lastSuccess:
          type: string
          format: date-time
        lastFailure:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    WebhookEvent:
      type: string
      enum:
        - agent.execution.started
        - agent.execution.completed
        - agent.execution.failed
        - workflow.started
        - workflow.completed
        - workflow.failed
        - approval.requested
        - approval.completed
        - approval.rejected
        - api_key.created
        - api_key.revoked
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not_found
              message:
                type: string
...